🚨 fix(test): 테스트 실패 원인 수정
