--- src/main/java/com/nextcloudlab/kickytime/controller/MatchingAdminController.java
@@ -1,24 +1,0 @@
-package com.nextcloudlab.kickytime.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.*;
-
-import com.nextcloudlab.kickytime.service.MatchingService;
-
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequestMapping("/api/matches")
-@RequiredArgsConstructor
-public class MatchingAdminController {
-
-    private final MatchingService matchingService;
-
-    @DeleteMapping("/{matchId}")
-    @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<Void> deleteMatch(@PathVariable Long matchId) {
-        matchingService.deleteMatchById(matchId);
-        return ResponseEntity.noContent().build();
-    }
-}
