+++ src/main/java/com/nextcloudlab/kickytime/match/controller/MatchingAdminController.java
@@ -1,0 +1,25 @@
+package com.nextcloudlab.kickytime.match.controller;
+
+import com.nextcloudlab.kickytime.match.service.MatchService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/matches")
+@RequiredArgsConstructor
+public class MatchingAdminController {
+
+    private final MatchService matchService;
+
+    @DeleteMapping("/{matchId}")
+    @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<Void> deleteMatch(@PathVariable Long matchId) {
+        matchService.deleteMatchById(matchId);
+        return ResponseEntity.noContent().build();
+    }
+}
